Move proxystore.p2p.counter to proxystore.utils
The AtomicCounter is a general utility class that is not specific to
proxystore.p2p.
