+++ proxystore/p2p/connection.py
@@ -32,11 +32,11 @@ except ImportError as e:  # pragma: no cover
 from proxystore.p2p.chunks import Chunk
 from proxystore.p2p.chunks import chunkify
 from proxystore.p2p.chunks import reconstruct
-from proxystore.p2p.counter import AtomicCounter
 from proxystore.p2p.exceptions import PeerConnectionError
 from proxystore.p2p.exceptions import PeerConnectionTimeoutError
 from proxystore.p2p.relay import RelayClient
 from proxystore.p2p.relay.messages import PeerConnectionRequest
+from proxystore.utils.counter import AtomicCounter
 
 logger = logging.getLogger(__name__)
